[PM-XXXX] Move i18n abstractions and services into new @bitwarden/i18n lib
Creates a dedicated `libs/i18n` package owned by the platform team.
Moves `TranslationService` and `I18nService` (abstractions + base implementations)
out of `libs/common` into the new lib. The original paths in `libs/common` are
replaced with deprecated re-exports so all existing consumers continue to work
without changes.
