+++ libs/common/src/platform/abstractions/i18n.service.ts
@@ -1,10 +1,2 @@
-import { Observable } from "rxjs";
-
-import { TranslationService } from "./translation.service";
-
-export abstract class I18nService extends TranslationService {
-  abstract userSetLocale$: Observable<string | undefined>;
-  abstract locale$: Observable<string>;
-  abstract setLocale(locale: string | null): Promise<void>;
-  abstract init(): Promise<void>;
-}
+/** @deprecated Import from `@bitwarden/i18n` instead */
+export { I18nServiceAbstraction as I18nService } from "@bitwarden/i18n";
