+++ libs/common/src/platform/abstractions/translation.service.ts
@@ -1,8 +1,2 @@
-export abstract class TranslationService {
-  abstract supportedTranslationLocales: string[];
-  abstract translationLocale: string;
-  abstract collator: Intl.Collator;
-  abstract localeNames: Map<string, string>;
-  abstract t(id: string, p1?: string | number, p2?: string | number, p3?: string | number): string;
-  abstract translate(id: string, p1?: string, p2?: string, p3?: string): string;
-}
+/** @deprecated Import from `@bitwarden/i18n` instead */
+export { TranslationServiceAbstraction as TranslationService } from "@bitwarden/i18n";
