+++ libs/i18n/README.md
@@ -1,0 +1,5 @@
+# i18n
+
+Owned by: platform
+
+Abstractions and base implementations for i18n
