+++ libs/i18n/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@bitwarden/i18n",
+  "version": "0.0.1",
+  "description": "Abstractions and base implementations for i18n",
+  "private": true,
+  "type": "commonjs",
+  "main": "index.js",
+  "types": "index.d.ts",
+  "license": "GPL-3.0",
+  "author": "platform"
+}
