+++ libs/i18n/src/abstractions/i18n.service.ts
@@ -1,0 +1,10 @@
+import { Observable } from "rxjs";
+
+import { TranslationService } from "./translation.service";
+
+export abstract class I18nService extends TranslationService {
+  abstract userSetLocale$: Observable<string | undefined>;
+  abstract locale$: Observable<string>;
+  abstract setLocale(locale: string | null): Promise<void>;
+  abstract init(): Promise<void>;
+}
