+++ libs/i18n/src/index.ts
@@ -1,0 +1,4 @@
+export { TranslationService as TranslationServiceAbstraction } from "./abstractions/translation.service";
+export { I18nService as I18nServiceAbstraction } from "./abstractions/i18n.service";
+export { TranslationService } from "./services/translation.service";
+export { I18nService } from "./services/i18n.service";
