+++ tsconfig.base.json
@@ -31,8 +31,8 @@
       "@bitwarden/browser/*": ["./apps/browser/src/*"],
       "@bitwarden/cli/*": ["./apps/cli/src/*"],
       "@bitwarden/client-type": ["./libs/client-type/src/index.ts"],
-      "@bitwarden/common/spec": ["./libs/common/spec"],
       "@bitwarden/common/*": ["./libs/common/src/*"],
+      "@bitwarden/common/spec": ["./libs/common/spec"],
       "@bitwarden/components": ["./libs/components/src"],
       "@bitwarden/core-test-utils": ["./libs/core-test-utils/src/index.ts"],
       "@bitwarden/dirt-card": ["./libs/dirt/card/src"],
@@ -42,6 +42,7 @@
       "@bitwarden/generator-legacy": ["./libs/tools/generator/extensions/legacy/src"],
       "@bitwarden/generator-navigation": ["./libs/tools/generator/extensions/navigation/src"],
       "@bitwarden/guid": ["./libs/guid/src/index.ts"],
+      "@bitwarden/i18n": ["./libs/i18n/src/index.ts"],
       "@bitwarden/importer-core": ["./libs/importer/src"],
       "@bitwarden/importer-ui": ["./libs/importer/src/components"],
       "@bitwarden/key-management": ["./libs/key-management/src"],
